Move package to separate module
